remove exclude from sampling, should be done via dataset.exclude_individuals

+++ src/automated_scoring/classification/utils.py
@@ -207,7 +207,7 @@ class EncodingFunction(Protocol):
         y: NDArray,
         *args,
         **kwargs,
-    ) -> NDArray[np.integer]: ...
+    ) -> NDArray[np.int64]: ...
 
 
 class SamplingFunction(Protocol):
